transfered data processing and rubric extraction functionalities to seperate file
