+++ pdll/_pairwise_NLP.py
@@ -1,129 +1,22 @@
-import codecs
 import os
-from pathlib import Path
-from sklearn.metrics import mean_squared_error
 
 import openai
 import pandas as pd
 from dotenv import load_dotenv
+from sklearn.metrics import mean_squared_error
+
+import pdll._pairwise_NLP_dataprocessing as data_processing
+import pdll._pairwise_NLP_rubricextraction as rubric_extraction
 
 load_dotenv()
 api_key = os.getenv("OPENAI_API_KEY")
 openai.api_key = api_key
 
 SEED = 17
-
-
-def get_rubric_texts_from_files(target_dir: Path) -> dict[int, str]:
-    file_dict = {}
-    for file in target_dir.iterdir():
-        if file.is_file() and file.suffix == ".txt":
-            essay_set_ID = int(file.stem.split("_")[2])
-
-            format_rubric_text_file(file, formatting_keywords, "Essay Set")
-
-            with open(file, "r", encoding="utf-8") as rf:
-                rubric_text = rf.read()
-
-            file_dict[essay_set_ID] = rubric_text
-    return file_dict
-
-
-def format_rubric_text_file(file: Path, keyword_list: list[str], filter: str):
-    with open(file, "r", encoding="utf-8") as reader, open(
-        file, "r+", encoding="utf-8"
-    ) as writer:
-        for line in reader:
-            if line.rstrip() and not line.startswith(filter):
-                for keyword in keyword_list:
-                    if line.startswith(keyword):
-                        writer.write("\n")
-                writer.write(line)
-        writer.truncate()
-
+FOLD_ID = 0
 
 # read rubric files
-rubric_dir = Path("data-set\\scoring_rubrics")
-formatting_keywords = [
-    "Prompt",
-    "Rubric Guidelines",
-    "Score",
-    "Adjudication Rules",
-    "Source Essay",
-    "Conventions",
-    "Sentence Fluency",
-    "Word Choice",
-    "Voice",
-    "Organization",
-    "Ideas",
-    "Style",
-    "Total Composite Score",
-]
-essay_set_ID_to_scoring_rubrics = get_rubric_texts_from_files(rubric_dir)
-
-
-# Input data file from corresponding fold
-FOLD_ID = 0
-dir = Path(f"data-set\\asap\\fold_{FOLD_ID}")
-paths = [f"{dir}\\train.tsv", f"{dir}\\dev.tsv", f"{dir}\\test.tsv"]
-prompt_id_to_score_index_mapping = {
-    1: 6,
-    2: 3,  #! todo: complete this mapping for all prompt ids
-    3: 3,
-    4: 3,
-    5: 3,
-    6: 3,
-    7: 3,
-}
-
-
-def read_dataset(file_path, prompt_id, score_index):
-    data_x, data_y, prompt_ids = [], [], []
-
-    with codecs.open(file_path, mode="r", encoding="UTF8") as input_file:
-        next(input_file)
-        for line in input_file:
-            tokens = line.strip().split("\t")
-
-            essay_id = int(tokens[0])
-            essay_set = int(tokens[1])
-            content = str(tokens[2].strip())
-            score = int(tokens[score_index])
-
-            if essay_set == prompt_id or prompt_id <= 0:
-                data_x.append(content)
-                data_y.append(score)
-                prompt_ids.append(essay_set)
-
-    return data_x, data_y, prompt_ids
-
-
-def get_data(paths, prompt_id, as_list_of_tuples):
-    train_path, dev_path, test_path = paths[0], paths[1], paths[2]
-
-    train_x, train_y, train_prompts = read_dataset(
-        train_path, prompt_id, prompt_id_to_score_index_mapping[prompt_id]
-    )
-
-    dev_x, dev_y, dev_prompts = read_dataset(
-        dev_path, prompt_id, prompt_id_to_score_index_mapping[prompt_id]
-    )
-
-    test_x, test_y, test_prompts = read_dataset(
-        test_path, prompt_id, prompt_id_to_score_index_mapping[prompt_id]
-    )
-
-    if as_list_of_tuples:
-        train = list(zip(train_x, train_y))
-        dev = list(zip(dev_x, dev_y))
-        test = list(zip(test_x, test_y))
-    else:
-        # as tuple of lists
-        train = (train_x, train_y)
-        dev = (dev_x, dev_y)
-        test = (test_x, test_y)
-
-    return train, dev, test
+essay_set_ID_to_scoring_rubrics = rubric_extraction.get_rubric_texts_from_files()
 
 
 def get_pair_diff_as_int(text1: str, text2: str, rubric: str) -> int:
@@ -165,20 +58,20 @@ def get_pair_diff_as_int(text1: str, text2: str, rubric: str) -> int:
         raise RuntimeError(f"OpenAI API call failed: {api_err}")
 
 
-# diff_1 = get_pair_diff_as_int(
-#     X_train.iloc[0].essay, X_train.iloc[1].essay, rubric_set_1_text
-# )
-# diff_2 = get_pair_diff_as_int(
-#     X_train.iloc[0].essay, X_train.iloc[1].essay, rubric_set_1_text
-# )
-
-# if diff_1 != diff_2:
-#     print(f"Different results! {diff_1} != {diff_2}")
-# else:
-#     print(f"Same results! Both say {diff_1}")
-
-# ? variation of the differences is high, so the model is not consistent in its predictions
-# ? range tested with this essay pair is -1 to 3, -1 x 1, 0 x 3, 1 x 2, 2 x 3, 3 x 1
+{
+    # diff_1 = get_pair_diff_as_int(
+    #     X_train.iloc[0].essay, X_train.iloc[1].essay, rubric_set_1_text
+    # )
+    # diff_2 = get_pair_diff_as_int(
+    #     X_train.iloc[0].essay, X_train.iloc[1].essay, rubric_set_1_text
+    # )
+    # if diff_1 != diff_2:
+    #     print(f"Different results! {diff_1} != {diff_2}")
+    # else:
+    #     print(f"Same results! Both say {diff_1}")
+    # ? variation of the differences is high, so the model is not consistent in its predictions
+    # ? range tested with this essay pair is -1 to 3, -1 x 1, 0 x 3, 1 x 2, 2 x 3, 3 x 1
+}  # type: ignore
 
 
 def predict_scores(test_data: pd.DataFrame, training_data: pd.DataFrame, rubric: str):
@@ -215,7 +108,9 @@ def predict_scores(test_data: pd.DataFrame, training_data: pd.DataFrame, rubric:
 
 essay_set_ID = 1  # Set the prompt ID you want to use
 as_list_of_tuples = True
-data_train, data_dev, data_test = get_data(paths, essay_set_ID, as_list_of_tuples)
+data_train, data_dev, data_test = data_processing.get_data(
+    FOLD_ID, essay_set_ID, as_list_of_tuples
+)
 
 limit = 6  # Limit the number of rows for testing
 limit_data = limit  # Limit the number of rows for testing
@@ -258,14 +153,14 @@ print(f"Mean Squared Error: {mse:.2f}")
 # include support for the other essay sets (other rubrics)
 # connect my work to the pre-folded / split data instead of the test-version
 
+# make sure the model doesn't have access to the internet so it doesn't just look-up;
+# api calls do not have native access to the internet
+
 # * TODO
 #! baseline first!
 # todo: with increasing complexity, bug detection might be more difficult
 # todo: include separation between modifications that should and shouldn't affect the score (for example cashing = no impact on score, prompt change = impact on score)
 # todo: implement a baseline predictor that only predicts the score directly from the model, for comparison
-
-# make sure the model doesn't have access to the internet so it doesn't just look-up;
-# api calls do not have native access to the internet
 
 # todo: cash get_pair_diff_as_int , reduces the API calls, use the FULL string, not only the inputs, because for example the prompt could be modified from one call to the other, cashing as dataframe, string and int diff for querying, printing to see which is a fresh API call and which come form the cash
